+++ lib/Controller/PageController.php
@@ -30,9 +30,9 @@ use OCA\Library\Http\ReviewQueryPolicy;
 use Throwable;
 
 class PageController extends Controller {
-    private const APP_VERSION = '0.1.0-alpha.168';
-    private const VUE_SCRIPT_ASSET = 'library-main-0-1-0-alpha-168-pathlink';
-    private const VUE_STYLE_ASSET = 'library-vue-0-1-0-alpha-168-pathlink';
+    private const APP_VERSION = '0.1.0-alpha.169';
+    private const VUE_SCRIPT_ASSET = 'library-main-0-1-0-alpha-169-pathlink';
+    private const VUE_STYLE_ASSET = 'library-vue-0-1-0-alpha-169-pathlink';
     private MonotonicClock $clock;
     /** @var array<string, true> */
     private array $invalidReviewKeys = [];